+++ changelog.md
@@ -1,6 +1,7 @@
 # Changelog
 
-- * e748abd - 2016-09-08: Fixed code in queue&#x27;s for handling currentItem 
+- * fd15622 - 2016-09-09: Fixed core queue code issues related to adding of shows Updated version to 8.1.1 
+- * 6ea6a06 - 2016-09-08: Fixed code in queue&#x27;s for handling currentItem 
 - * a2e97e8 - 2016-09-08: Fixed code in queue&#x27;s for handling currentItem 
 - * 3904a3c - 2016-09-08: Fixed code in queue&#x27;s for handling currentItem 
 - * dfa5e99 - 2016-09-08: Fixed code in queue&#x27;s for handling currentItem 
@@ -12,5 +13,4 @@
 - * e4a8a8a - 2016-09-04: Fixed KeyError for get_size function 
 - * 5494ed7 - 2016-09-03: Disabled raising exceptions for getting file sizes during provider searches 
 - * cd19664 - 2016-09-02: Fixed issues with getting overall size of search results. Added code to append list of files and sizes relating to the search results from the providers. Added code for quality definition video file size constaint checking. 
-- * 321616e - 2016-09-01: Fixed issues with core queue code